asm: split out AST into separate package
This change introduces a new package internal/ast for the AST and parser. Before, the
compiler used the AST to store its internal state, and even defined methods directly on
AST objects. As I move towards the implementation of a gofmt-like tool and stack checker,
I need an intermediate representation that can be shared across packages.

The compiler now has a concept of 'sections' in the output program. At this time, sections
are used to store a reference to the AST document and macro arguments for use by the
evaluator. They do not influence the output in any other way. However, the concept of
sections will also be useful for the implementation of EOF support.

Note that the Go compiler version requirement is updated to Go 1.23.