asm, internal/ast: require = in expression macro definitions (#12)
This is a language change. The syntax for expression macro definitions had an annoying
ambiguity - the parser could not distinguish the parameter list from a parenthesized body
expression. Definitions below didn't compile even though they look valid:

	#define macro1 (1)                
	#define macro2 (100 + a) / b


I noticed this quite late in the development process of the initial release, and didn't
have time to fix it. So I'm fixing it now. In the new syntax, you have to put `=` before
the macro body expression.

	;; new definition syntax
	#define macro1 = 1
	#define macro2 = (100 + a) / b
	#define macro3(p) = $p + 1

Legacy definition syntax is still allowed, but will produce a warning.

Aside from fixing the grammar ambiguity issue, the new syntax also just feels better. It's
a subjective thing, but for me, it is somehow easier to parse definitions mentally with
the separator - especially for small values. You can also use the `=` to align definitions
with each other. It was possible before, you could align the values, but it looked kind of
awkward.