asm: refactor compilerProg and label assignment
This is a collection of refactoring changes, improving the internal program data
structure and documentation. compilerProg has changed to hold a flat list of
instructions instead of nested section objects. It's much simpler to iterate this
way, and faster too. '#bytes' instructions are now handled explicitly, instead of
emitting empty instructions for them.

For labels, the compiler now tracks the pointed-to instruction instead of their PC.
This allows reasoning about labels directly after the expansion phase, and it seems
more correct in general.