asm: allow referencing builtin macros without the dot
This change is meant to simplify the language, especially for casual users.
While the expression language provides a facility to create macros with
parameters, it is rarely used by real programs. Most macro calls with parameters
are to simple builtins like len() or selector().

Given this observation, I feel it isn't appropriate forcing people to use dotted
syntax to access builtins. I added the dotted syntax to avoid any possible name
collisions of user-defined macros with user-defined ones - the dot effectively
creates a separate namespace. Builtins can still be referred to by their dotted
name, the change simply makes it possible to also refer to them without.

The Go language also has builtins, and the way they are handled has once again
served as an inspiration here. Redefining builtins is allowed for the simple
reason that existing programs might contain definitions with the same name. This
is especially important since the set of builtins may grow in the future.